Extract light conditions module
